+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This file is used to list changes made in each version of the openstack-block-st
 ## 10.1.0
 * Add disable logic for stage fix to tgtd issue on RHEL 7
 * Use common specific_endpoint routines (bug 1412919)
+* Install qemu-img package for volume nodes
 
 ## 10.0.1
 # Update cinder.conf mode from 0644 to 0640